requirements.txt
duckdb
geopandas
s3fs
pyarrow
openpyxl
tqdm